Move generateRandomTileMap to a utility class for generating maps
Related to #4

Move the `generateRandomTilemap` method to a utility class and rename it to `generateMap`.

* Add a new file `src/utils/MapGenerator.ts` with a `MapGenerator` class containing a static `generateMap` method.
* Move the logic from `generateRandomTilemap` to the `generateMap` method in the `MapGenerator` class.
* Modify `src/scenes/PlayScene.ts` to import the `MapGenerator` class and use the `generateMap` method instead of `generateRandomTilemap`.
* Remove the `generateRandomTilemap` method from the `PlayScene` class.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/abrie/ai-game-template/issues/4?shareId=XXXX-XXXX-XXXX-XXXX).
